requirements.txt
easydict
numpy
pybullet
imageio
pycocotools
pillow
pyyaml